# openapi.yml
openapi: 3.1.0
info:
  title: AgoraLume API
  description: |-
    User + multi-AI group chat with modular personas. The backend owns the workspace and streams chat over SSE.

    POST creates (201), PATCH merges a partial body (200), DELETE returns 204. Failures are RFC 9457 problem documents; `type` is a stable `urn:agoralume:error:…`.
  license:
    name: AGPL-3.0-only
    identifier: AGPL-3.0-only
  version: 0.1.0
servers:
- url: /
  description: 'Same-origin: the bundled build, where the SPA and the API share a host'
- url: /api
  description: Behind an edge/dev proxy (VITE_API_PREFIX=/api)
- url: http://127.0.0.1:8080
  description: A local backend (pnpm dev:api)
paths:
...
  /v1beta/llm/models:
    post:
      tags:
      - llm
      summary: |-
        Lists the models a provider endpoint offers, so the model field can be a
        picker. Admin-only. `apiKey` is required unless `baseUrl` matches the
        already-configured endpoint.
      operationId: list_llm_models
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LlmModelsQuery'
        required: true
      responses:
        '200':
          description: The models that endpoint reports
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmModelsView'
        '401':
          description: Not signed in
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Not permitted for this identity
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Empty or non-HTTP(S) baseUrl, no usable key, or the endpoint refused
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1beta/llm/settings:
    get:
      tags:
      - llm
      summary: |-
        The live LLM provider configuration. Any signed-in caller may read it;
        `canEdit` says whether this caller may write it.
      operationId: get_llm_settings
      responses:
        '200':
          description: The live configuration; the API key is reduced to `hasApiKey`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmSettingsView'
        '401':
          description: Not signed in
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Not permitted for this identity
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
    patch:
      tags:
      - llm
      summary: |-
        Merges a partial update onto the LLM provider configuration and applies it
        immediately. Admin-only.
      operationId: update_llm_settings
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LlmSettingsPatch'
        required: true
      responses:
        '200':
          description: Applied and persisted to llm.toml; no restart needed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LlmSettingsView'
        '401':
          description: Not signed in
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '403':
          description: Not permitted for this identity
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: The resulting configuration would not build; nothing changed
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
...
    ApiError:
      type: object
      description: |-
        An RFC 9457 problem document — the body of every 4xx this API produces.

        Constructed through the named helpers below rather than field-by-field, so
        the slug used for a given status is decided in exactly one place and can't
        drift between handlers.
      required:
      - type
      - title
      - status
      properties:
        detail:
          type:
          - string
          - 'null'
          description: |-
            The human-readable explanation of *this specific* occurrence — the
            sentence that used to be the entire `text/plain` body. Safe to show a
            user; never carries a secret, a stack trace, or a provider's raw
            response.
        status:
          type: integer
          format: int32
          description: |-
            The HTTP status code, repeated in the body so a problem document stays
            meaningful when it's logged or forwarded away from its response.
          minimum: 0
        title:
          type: string
          description: |-
            The status code's canonical reason phrase (e.g. `Forbidden`) — a short
            human summary of the problem *type*, not of this occurrence.
        type:
          type: string
          description: |-
            Stable, machine-readable identifier for *what* went wrong, e.g.
            `urn:agoralume:error:not-admin`. Match on this rather than on `detail`,
            whose wording is free to change. Always present.
          example: urn:agoralume:error:not-found
...
    LlmModelInfo:
      type: object
      description: |-
        One model a provider's listing endpoint reported. `name` is a display label
        when the provider offers one (Gemini does; OpenAI-compatible listings
        usually don't), never required for choosing the model — `id` is what's
        actually sent as `LlmSettings::model`.
      required:
      - id
      properties:
        id:
          type: string
        name:
          type:
          - string
          - 'null'
    LlmModelsQuery:
      type: object
      description: |-
        The `POST /llm/models` request: which endpoint to list models from. `apiKey`
        is optional — omitted, the backend uses the currently-stored key, but only
        when `baseUrl` matches the stored endpoint (see `routes::llm`), so this
        can't be used to make the server send its stored credential to an arbitrary
        third-party URL.
      required:
      - baseUrl
      properties:
        apiKey:
          type:
          - string
          - 'null'
        baseUrl:
          type: string
    LlmModelsView:
      type: object
      description: The `POST /llm/models` response.
      required:
      - models
      properties:
        models:
          type: array
          items:
            $ref: '#/components/schemas/LlmModelInfo'
    LlmSettingsPatch:
      type: object
      description: |-
        A partial update to the LLM provider configuration — every field is
        optional; an absent field leaves the current value alone. `baseUrl` /
        `model` / `apiKey` treat an empty string as "clear this field", matching the
        endpoint's other optional-string convention (there's never a reason to
        *store* an empty string here). `pricing` with both rates at zero clears the
        configured pricing (shows token counts only) rather than pinning a $0 rate.
      properties:
        apiKey:
          type:
          - string
          - 'null'
          description: |-
            The new key, or `""` to clear it. Omit entirely to leave the stored key
            untouched — the frontend must never send this unless the operator
            actually typed a new value, or every unrelated settings save would wipe
            the key.
        baseUrl:
          type:
          - string
          - 'null'
        compressAfter:
          type:
          - integer
          - 'null'
          minimum: 0
        compressKeep:
          type:
          - integer
          - 'null'
          minimum: 0
        compressMaxTokens:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
        enabled:
          type:
          - boolean
          - 'null'
        maxRetries:
          type:
          - integer
          - 'null'
          format: int32
          minimum: 0
        maxRpm:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
        maxTokens:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
        model:
          type:
          - string
          - 'null'
        pricing:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Pricing'
        retryBaseMs:
          type:
          - integer
          - 'null'
          format: int64
          minimum: 0
    LlmSettingsView:
      type: object
      description: |-
        The `GET`/`PATCH /llm/settings` response: the live LLM provider
        configuration, with the API key stripped to a presence flag. Built by hand
        from [`crate::llm_config::LlmSettings`] (never derived by serializing it
        directly) — that's the one place a leaked key would slip out.
      required:
      - enabled
      - hasApiKey
      - maxTokens
      - maxRpm
      - maxRetries
      - retryBaseMs
      - compressAfter
      - compressKeep
      - compressMaxTokens
      - canEdit
      properties:
        baseUrl:
          type:
          - string
          - 'null'
        canEdit:
          type: boolean
          description: |-
            Whether *this caller* may write this config (`PATCH /llm/settings`,
            `POST /llm/models`) — admin-only today (see `CurrentAdmin` in
            `backend/src/state.rs`), computed by the `GET /llm/settings` handler
            from the caller's actual `Subject`, not by anything in this struct's
            `From` impl (which always seeds it `false`, having no caller to ask).
            The frontend keys its write controls off this field instead of
            re-deriving "am I admin" from its own copy of the session role, so it
            can't drift from whatever the server actually enforces.
        compressAfter:
          type: integer
          minimum: 0
        compressKeep:
          type: integer
          minimum: 0
        compressMaxTokens:
          type: integer
          format: int64
          minimum: 0
        enabled:
          type: boolean
        hasApiKey:
          type: boolean
          description: |-
            Whether a key is currently stored. The key itself is never sent to a
            client once saved.
        maxRetries:
          type: integer
          format: int32
          minimum: 0
        maxRpm:
          type: integer
          format: int64
          minimum: 0
        maxTokens:
          type: integer
          format: int64
          minimum: 0
        model:
          type:
          - string
          - 'null'
        pricing:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/Pricing'
        retryBaseMs:
          type: integer
          format: int64
          minimum: 0
...
    Pricing:
      type: object
      description: |-
        Token pricing used to turn usage into an estimated cost. Rates are per
        1,000,000 tokens, in `currency`. A rough operator-supplied estimate — models
        and providers differ — so the UI always labels the result "for reference".
        Doubles as a wire type (`GET`/`PATCH /llm/settings`), hence `camelCase`; the
        one place that reads a little unusual is its own corner of `llm.toml`.
      required:
      - inputPerM
      - cachedInputPerM
      - outputPerM
      - currency
      properties:
        cachedInputPerM:
          type: number
          format: double
          description: |-
            Price per 1M cached input tokens. Usually cheaper than fresh input;
            defaults to the fresh input rate when not set (a conservative estimate
            that shows no cache saving until a discounted rate is provided).
        currency:
          type: string
          description: Currency label the rates are quoted in.
        inputPerM:
          type: number
          format: double
          description: Price per 1M fresh (non-cached) input tokens.
        outputPerM:
          type: number
          format: double
          description: Price per 1M output tokens.
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 'Access token from `POST /auth/login`. Opaque; does not survive a server restart. Waived when `GET /meta` reports `authRequired: false`.'
security:
- bearerAuth: []
...